requirements.txt
streamlit
groq
chromadb
langchain-groq
pdfplumber
pymupdf
mysql-connector-python
python-dotenv
pysqlite3-binary